çözüm yapıldı fakat optimum değil

+++ src/antcolony/Ant.java
@@ -15,9 +15,9 @@ import jdk.nashorn.internal.runtime.arrays.ArrayLikeIterator;
 public class Ant {
     
   public String name;
-   public ArrayList<Boolean> visited ;//town index -> true or false
+    public ArrayList<Boolean> visited ;//town index -> true or false
     public ArrayList<Town> tours; // tour list for each ants 
-    public int tourLength;
+    public long tourLength;
     public String bestTour; // best solition from tour list for one ant
     public Town currentTown;
     public int currentIndex;
